# improve ctor and dispose